# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  category: string;
  price: number;
  description: string;
  imageUrl: string;
  tags: string[];
  featured: boolean;
  available: boolean;
  createdAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  category: string;
  /** @minimum 0 */
  price: number;
  /** @minLength 1 */
  description: string;
  /** @minLength 1 */
  imageUrl: string;
  tags?: string[];
  featured?: boolean;
  available?: boolean;
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  slug?: string;
  /** @minLength 1 */
  category?: string;
  /** @minimum 0 */
  price?: number;
  /** @minLength 1 */
  description?: string;
  /** @minLength 1 */
  imageUrl?: string;
  tags?: string[];
  featured?: boolean;
  available?: boolean;
}

export interface CategorySummary {
  category: string;
  count: number;
  imageUrl: string;
}

export interface StoreSummary {
  productCount: number;
  categoryCount: number;
  featuredCount: number;
}

export type ListProductsParams = {
category?: string;
search?: string;
featured?: boolean;
available?: boolean;
};