# lib/api-zod/src/generated/types/categorySummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CategorySummary {
  category: string;
  count: number;
  imageUrl: string;
}