# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  category: string;
  /** @minimum 0 */
  price: number;
  /** @minLength 1 */
  description: string;
  /** @minLength 1 */
  imageUrl: string;
  tags?: string[];
  featured?: boolean;
  available?: boolean;
}